security(sqlite): bump bundled SQLite to 3.51.1 (>= 3.50.3 security floor)
Security-floor upgrade: rusqlite 0.32 -> 0.38, libsqlite3-sys 0.30.1 ->
0.36.0. The `bundled` feature compiles plain sqlite3 3.51.1 (no sqlcipher
variant is used), past the 3.50.3 security floor.

CVEs fixed by this upgrade:
- CVE-2025-7709 (3.50.3): corrupt FTS5 index -> memory access
- CVE-2025-6965 (3.50.2): aggregate term count overflow
- CVE-2025-29087 / -3277 (3.49.1): concat_ws() integer overflow
- CVE-2025-29088 (3.49.1): lookaside allocator DoS (separate from concat)

CVE-2025-7709 directly threatens memory-core: FTS5 is the core search
mechanism and DB content is user-writable (memories, wiki, events, URL
ingest). Refs #833.

API adaptation (rusqlite 0.38 removed usize FromSql/ToSql):
- daily_pipeline.rs, memory_crud/read.rs, store/maintenance.rs,
  bootstrap/tidy/migration.rs: usize -> i64 at SQL boundary (SQLite has
  no native unsigned type; i64 is the canonical mapping)

Test adaptation (genuine SQLite 3.50+ behavior change):
- doctor_probe_ops schema_version_fails_on_garbage_bytes: on SQLite
  >=3.50 the "not a database" error surfaces at open time when an
  auto-extension (libsimple) is registered, not at query time. Adapted
  the assertion to accept the error at either step — does not weaken the
  intent (garbage file must still error).

New security-floor assertion tests (memory-core sqlite_security_floor):
- bundled_sqlite_meets_security_floor_3_50_3 (CVE-2025-7709 reference)
- bundled_sqlite_has_fts5
- bundled_sqlite_has_json1

Verified:
- rusqlite 0.38.0 / libsqlite3-sys 0.36.0 / bundled sqlite3 3.51.1
- cargo test -p memory-core --lib: 265 passed; 0 failed
- cargo test -p memory-server --lib: 1341 passed; 0 failed
- cargo clippy -p memory-core -p memory-server -- -D warnings: clean
- cargo audit: exit 0 (note: cargo audit cannot detect bundled C SQLite
  CVEs; the assertion test is the floor enforcement)

+++ crates/memory-core/src/db/memory_crud/read.rs
@@ -107,7 +107,7 @@ pub fn get_all(
         )
     };
     let mut stmt = conn.prepare(&sql)?;
-    let rows = stmt.query_map(params![limit], row_to_entry)?;
+    let rows = stmt.query_map(params![limit as i64], row_to_entry)?;
 
     let mut out = Vec::new();
     for r in rows {
@@ -158,7 +158,7 @@ pub fn list_by_path(
     };
 
     let mut stmt = conn.prepare(&sql)?;
-    let rows = stmt.query_map(params![normalized, like_prefix, limit], row_to_entry)?;
+    let rows = stmt.query_map(params![normalized, like_prefix, limit as i64], row_to_entry)?;
     let mut out = Vec::new();
     for r in rows {
         out.push(r?);
@@ -188,7 +188,7 @@ pub fn list_wiki_duplicate_candidates(
     );
     let mut stmt = conn.prepare(&sql)?;
     let rows = stmt.query_map(
-        params![path, topic, parent_path, parent_like, limit],
+        params![path, topic, parent_path, parent_like, limit as i64],
         row_to_entry,
     )?;
     let mut out = Vec::new();
